hrgsms-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
mysql-connector-python>=9.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
pydantic-settings>=2.2.1
